requirements.txt
Django==3.0.1
django-filter==2.2.0
djangorestframework==3.11.0
flake8==3.7.9
gunicorn==20.0.4
Markdown==3.1.1
psycopg2==2.8.4
ptvsd==4.3.2
PyJWT==1.7.1